+++ MatrixKit/Models/Account/MXKAccount.m
@@ -113,6 +113,9 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
         [self prepareRESTClient];
         
         userPresence = MXPresenceUnknown;
+        
+        // Refresh device information
+        [self loadDeviceInformation:nil failure:nil];
     }
     return self;
 }
@@ -153,6 +156,11 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
         {
             threePIDs = [coder decodeObjectForKey:@"threePIDs"];
         }
+        
+        if ([coder decodeObjectForKey:@"device"])
+        {
+            _device = [coder decodeObjectForKey:@"device"];
+        }
 
         userPresence = MXPresenceUnknown;
         
@@ -175,6 +183,9 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
         _enableInAppNotifications = [coder decodeBoolForKey:@"enableInAppNotifications"];
         
         _disabled = [coder decodeBoolForKey:@"disabled"];
+        
+        // Refresh device information
+        [self loadDeviceInformation:nil failure:nil];
     }
     
     return self;
@@ -199,6 +210,11 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
     if (self.threePIDs)
     {
         [coder encodeObject:threePIDs forKey:@"threePIDs"];
+    }
+    
+    if (self.device)
+    {
+        [coder encodeObject:_device forKey:@"device"];
     }
 
     if (self.identityServerURL)
@@ -456,6 +472,41 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
             failure(error);
         }
     }];
+}
+
+- (void)loadDeviceInformation:(void (^)())success failure:(void (^)(NSError *error))failure
+{
+    if (mxCredentials.deviceId)
+    {
+        [mxRestClient deviceByDeviceId:mxCredentials.deviceId success:^(MXDevice *device) {
+            
+            _device = device;
+            
+            // Archive updated field
+            [[MXKAccountManager sharedManager] saveAccounts];
+            
+            if (success)
+            {
+                success();
+            }
+            
+        } failure:^(NSError *error) {
+            
+            if (failure)
+            {
+                failure(error);
+            }
+            
+        }];
+    }
+    else
+    {
+        _device = nil;
+        if (success)
+        {
+            success();
+        }
+    }
 }
 
 - (void)setUserPresence:(MXPresence)presence andStatusMessage:(NSString *)statusMessage completion:(void (^)(void))completion