+++ MatrixKit/Views/RoomInputToolbar/MXKRoomInputToolbarViewWithSimpleTextView.m
@@ -74,14 +74,6 @@
 
 #pragma mark - UITextViewDelegate
 
-- (void)textViewDidBeginEditing:(UITextView *)textView
-{
-    if ([self.delegate respondsToSelector:@selector(roomInputToolbarView:isTyping:)])
-    {
-        [self.delegate roomInputToolbarView:self isTyping:YES];
-    }
-}
-
 - (void)textViewDidEndEditing:(UITextView *)textView
 {
     if ([self.delegate respondsToSelector:@selector(roomInputToolbarView:isTyping:)])